+++ data/lib/us_geo_data/zcta.rb
@@ -6,11 +6,12 @@ module USGeoData
 
     def dump_csv(output)
       csv = CSV.new(output)
-      csv << ["ZCTA5", "Primary County", "Primary Place", "Population", "Housing Units", "Land Area", "Water Area", "Latitude", "Longitude"]
+      csv << ["ZCTA5", "Primary County", "Primary County Subdivision", "Primary Place", "Population", "Housing Units", "Land Area", "Water Area", "Latitude", "Longitude"]
       zcta_data.each_value do |data|
         csv << [
           data[:zcta],
           data[:primary_county],
+          data[:primary_county_subdivision],
           data[:primary_place],
           data[:population],
           data[:housing_units],
@@ -31,8 +32,24 @@ module USGeoData
           csv << [
             zcta[:zcta],
             county_geoid,
-            area[:land_area],
-            area[:water_area]
+            area[:land_area]&.round(3),
+            area[:water_area]&.round(3)
+          ]
+        end
+      end
+      output
+    end
+
+    def dump_county_subdivisions_csv(output)
+      csv = CSV.new(output)
+      csv << ["ZCTA5", "County Subdivision GEOID", "Land Area", "Water Area"]
+      zcta_data.each_value do |zcta|
+        zcta[:county_subdivisions].each do |county_subdivision_geoid, area|
+          csv << [
+            zcta[:zcta],
+            county_subdivision_geoid,
+            area[:land_area]&.round(3),
+            area[:water_area]&.round(3)
           ]
         end
       end
@@ -79,6 +96,7 @@ module USGeoData
         end
 
         add_counties(data)
+        add_county_subdivisions(data)
         add_places(data)
         add_demographics(data)
 
@@ -132,12 +150,12 @@ module USGeoData
       foreach(data_file(USGeoData::ZCTA_COUNTY_REL_FILE), col_sep: "|") do |row|
         zcta5 = row["GEOID_ZCTA5_20"]
         county_geoid = row["GEOID_COUNTY_20"]
-        county_land_area = row["AREALAND_PART"].to_f * SQUARE_METERS_TO_MILES
-        county_water_area = row["AREAWATER_PART"].to_f * SQUARE_METERS_TO_MILES
-        next unless zcta5 && county_geoid && county_land_area > 0
+        overlap_land_area = row["AREALAND_PART"].to_f * SQUARE_METERS_TO_MILES
+        overlap_water_area = row["AREAWATER_PART"].to_f * SQUARE_METERS_TO_MILES
+        next unless zcta5 && county_geoid && overlap_land_area > 0
 
         info = data[zcta5]
-        info[:counties][county_geoid] = {land_area: county_land_area, water_area: county_water_area}
+        info[:counties][county_geoid] = {land_area: overlap_land_area, water_area: overlap_water_area}
       end
 
       data.each_value do |info|
@@ -145,18 +163,35 @@ module USGeoData
       end
     end
 
+    def add_county_subdivisions(data)
+      foreach(data_file(USGeoData::ZCTA_COUNTY_SUBDIVISION_REL_FILE), col_sep: "|") do |row|
+        zcta5 = row["GEOID_ZCTA5_20"]
+        county_subdivision_geoid = row["GEOID_COUSUB_20"]
+        overlap_land_area = row["AREALAND_PART"].to_f * SQUARE_METERS_TO_MILES
+        overlap_water_area = row["AREAWATER_PART"].to_f * SQUARE_METERS_TO_MILES
+        next unless zcta5 && county_subdivision_geoid && overlap_land_area > 0
+
+        info = data[zcta5]
+        info[:county_subdivisions][county_subdivision_geoid] = {land_area: overlap_land_area, water_area: overlap_water_area}
+      end
+
+      data.each_value do |info|
+        info[:primary_county_subdivision] = info[:county_subdivisions].max_by { |_, area| area[:land_area] }&.first
+      end
+    end
+
     def add_places(data)
       foreach(data_file(USGeoData::ZCTA_PLACE_REL_FILE), col_sep: "|") do |row|
         zcta5 = row["GEOID_ZCTA5_20"]
         place_geoid = row["GEOID_PLACE_20"]
-        place_land_area = row["AREALAND_PART"].to_f * SQUARE_METERS_TO_MILES
-        place_water_area = row["AREAWATER_PART"].to_f * SQUARE_METERS_TO_MILES
-        next unless place_geoid && place_land_area > 0
+        overlap_land_area = row["AREALAND_PART"].to_f * SQUARE_METERS_TO_MILES
+        overlap_water_area = row["AREAWATER_PART"].to_f * SQUARE_METERS_TO_MILES
+        next unless place_geoid && overlap_land_area > 0
 
         info = data[zcta5]
         next unless info
 
-        info[:places][place_geoid] = {land_area: place_land_area, water_area: place_water_area}
+        info[:places][place_geoid] = {land_area: overlap_land_area, water_area: overlap_water_area}
       end
 
       data.each_value do |info|
@@ -172,6 +207,7 @@ module USGeoData
         land_area: 0.0,
         water_area: 0.0,
         counties: {},
+        county_subdivisions: {},
         places: {}
       }
     end